Move some test cases to proper place
